requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.9.0
ccxt==4.4.28
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.0.2
pandas==2.2.3
propcache==0.2.0
psycopg2-binary==2.9.10
pycares==4.4.0
pycparser==2.22
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==2.3.7
yarl==1.17.1
asgiref-3.8.1